requirements/dev38.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/dev38.txt requirements/dev.in
#
amqp==5.0.5
    # via kombu
appdirs==1.4.4
    # via virtualenv
appnope==0.1.2
    # via ipython
argh==0.26.2
    # via -r requirements/dev.in
asgiref==3.3.1
    # via django
attrs==20.3.0
    # via pytest
backcall==0.2.0
    # via ipython
billiard==3.6.3.0
    # via celery
bleach==3.3.0
    # via readme-renderer
bump2version==1.0.1
    # via bumpversion
bumpversion==0.6.0
    # via -r requirements/dev.in
celery[redis]==5.0.5
    # via
    #   -r requirements/dev.in
    #   django-celery-beat
    #   django-celery-results
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.1.6
    # via celery
click==7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
colorama==0.4.4
    # via twine
coverage==5.5
    # via pytest-cov
decorator==4.4.2
    # via ipython
distlib==0.3.1
    # via virtualenv
django-celery-beat==2.2.0
    # via -r requirements/dev.in
django-celery-results==2.0.1
    # via -r requirements/dev.in
django-debug-toolbar==3.2
    # via -r requirements/dev.in
django-environ==0.4.5
    # via -r requirements/dev.in
django-extensions==3.1.1
    # via -r requirements/dev.in
django-redis==4.12.1
    # via -r requirements/dev.in
django-site-utils==0.3.7
    # via -r requirements/dev.in
django-timezone-field==4.1.2
    # via django-celery-beat
django==3.1.7
    # via
    #   -r requirements/dev.in
    #   django-celery-beat
    #   django-debug-toolbar
    #   django-redis
    #   django-site-utils
    #   django-timezone-field
docutils==0.16
    # via readme-renderer
filelock==3.0.12
    # via
    #   tox
    #   virtualenv
flake8==3.9.0
    # via
    #   -r requirements/dev.in
    #   pytest-flake8
honcho==1.0.1
    # via -r requirements/dev.in
idna==2.10
    # via requests
importlib-metadata==3.7.3
    # via
    #   keyring
    #   twine
iniconfig==1.1.1
    # via pytest
ipython-genutils==0.2.0
    # via traitlets
ipython==7.21.0
    # via -r requirements/dev.in
jedi==0.18.0
    # via ipython
keyring==23.0.0
    # via twine
kombu==5.0.2
    # via celery
mccabe==0.6.1
    # via flake8
packaging==20.9
    # via
    #   bleach
    #   pytest
    #   tox
parso==0.8.1
    # via jedi
pep517==0.10.0
    # via pip-tools
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==6.0.1
    # via -r requirements/dev.in
pkginfo==1.7.0
    # via twine
pluggy==0.13.1
    # via
    #   pytest
    #   tox
prompt-toolkit==3.0.17
    # via
    #   click-repl
    #   ipython
psycopg2-binary==2.8.6
    # via -r requirements/dev.in
ptyprocess==0.7.0
    # via pexpect
py==1.10.0
    # via
    #   pytest
    #   tox
pycodestyle==2.7.0
    # via
    #   -r requirements/dev.in
    #   flake8
pyflakes==2.3.0
    # via flake8
pygments==2.8.1
    # via
    #   ipython
    #   readme-renderer
pyparsing==2.4.7
    # via packaging
pytest-cov==2.11.1
    # via -r requirements/dev.in
pytest-django==4.1.0
    # via -r requirements/dev.in
pytest-flake8==1.0.7
    # via -r requirements/dev.in
pytest-runner==5.3.0
    # via -r requirements/dev.in
pytest==6.2.2
    # via
    #   -r requirements/dev.in
    #   pytest-cov
    #   pytest-django
    #   pytest-flake8
python-crontab==2.5.1
    # via django-celery-beat
python-dateutil==2.8.1
    # via python-crontab
python-redis-lock==3.7.0
    # via -r requirements/dev.in
pytz==2021.1
    # via
    #   celery
    #   django
    #   django-timezone-field
pyyaml==5.4.1
    # via -r requirements/dev.in
readme-renderer==29.0
    # via twine
redis==3.5.3
    # via
    #   celery
    #   django-redis
    #   python-redis-lock
requests-toolbelt==0.9.1
    # via twine
requests==2.25.1
    # via
    #   requests-toolbelt
    #   twine
rfc3986==1.4.0
    # via twine
setuptools-twine==0.1.3
    # via -r requirements/dev.in
six==1.15.0
    # via
    #   bleach
    #   click-repl
    #   django-site-utils
    #   python-dateutil
    #   readme-renderer
    #   tox
    #   virtualenv
sqlparse==0.4.1
    # via
    #   django
    #   django-debug-toolbar
toml==0.10.2
    # via
    #   pep517
    #   pytest
    #   tox
tox-gh-actions==2.4.0
    # via -r requirements/dev.in
tox==3.23.0
    # via
    #   -r requirements/dev.in
    #   tox-gh-actions
tqdm==4.59.0
    # via twine
traitlets==5.0.5
    # via ipython
twine==3.4.1
    # via
    #   -r requirements/dev.in
    #   setuptools-twine
urllib3==1.26.4
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
virtualenv==20.4.3
    # via tox
watchdog==2.0.2
    # via -r requirements/dev.in
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via bleach
zipp==3.4.1
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
